# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/{projectId}/teams": {
      "post": {
        "operationId": "TeamsController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "에이전트 팀 생성",
        "tags": [
          "Teams"
        ]
      },
      "get": {
        "operationId": "TeamsController_list",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "에이전트 팀 목록 조회",
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/teams/{teamId}": {
      "get": {
        "operationId": "TeamDetailController_get",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "팀 상세 조회",
        "tags": [
          "Teams"
        ]
      },
      "patch": {
        "operationId": "TeamDetailController_update",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "팀 수정",
        "tags": [
          "Teams"
        ]
      },
      "delete": {
        "operationId": "TeamDetailController_delete",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "팀 삭제",
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/teams/{teamId}/agents": {
      "post": {
        "operationId": "TeamDetailController_addAgent",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "팀에 에이전트 추가",
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/teams/{teamId}/agents/{agentId}": {
      "delete": {
        "operationId": "TeamDetailController_removeAgent",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "팀에서 에이전트 제거",
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/teams/{teamId}/subteams": {
      "post": {
        "operationId": "TeamDetailController_addSubTeam",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "팀에 서브팀 추가 (Nested Team)",
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/teams/{teamId}/subteams/{subTeamId}": {
      "delete": {
        "operationId": "TeamDetailController_removeSubTeam",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subTeamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "팀에서 서브팀 제거",
        "tags": [
          "Teams"
        ]
      }
    },
    "/api/teams/{teamId}/invoke": {
      "post": {
        "operationId": "TeamDetailController_invoke",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "팀 실행 (Playground 용 stub)",
        "tags": [
          "Teams"
        ]
      }
    },
...
  },
  "info": {
    "title": "AgentStudio API",
    "description": "AgentStudio 관리 API. Client UI(`apps/agent-web`) 페이지별 사용 매핑은 `API.md` §10 참조.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4200",
      "description": "local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "POST /api/auth/login 으로 발급받은 accessToken"
      }
    },
...
  }
}